requirements.txt
pandas==2.2.0
numpy==1.26.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
